new-york/requirements.txt
openai 
psycopg2-binary
requests
bs4
python-docx
twilio
python-dotenv
s